Commiting work already done.
So far there are three modules included in this repository:

- git
- julia-from-source
- update-deb-sources

The `git` module currently only contains code to install git with a
package manager and to clone a repository from github.

The `julia-from-source` module contains code to install the Julia
Language, by first installing its dependencies, then fetching its
source from github and running `make` and `make install`. This is
currently a very long process, since `make` takes forever to run.

The `update-deb-sources` is a package that when run, will run the
command `apt-get update`. This is convenient to have in a module,
since it is usually nice to update the repositories before trying to
install any software.