Finished emacs module, fixed bugs in git, update.
The emacs version can now successfully install emacs and download a
emacs configuration from github, putting it in ~/.emacs.d.

The bug fixes are fairly numerous.

  1) In the `git` module, we now insure that git is installed when
     trying to call `git::github-clone`.

  2) Adding the `$name` parameter in the the `git::github-clone`
     definition introduced a bug where if that directory existed, the
     code would exit successfully without cloning the repo. This bug
     has been fixed.

  3) In the `update-repos` module there was a bug with how the
     `$disto` variable was being used, this is fixed.

  4) The `update-repos` module no longer incorrectly warns that you're
     installing debian sources.list in ubuntu and vice versa.

Note: The solution to bug (2) is to delete the directory under `$name`
if it exists, which is somewhat reasonable. However, a better solution
might be to fail if this directory already exists.

------------------------------------------------------------------------