Use the anchor pattern so updates occur in order.
Classes declared within classes can float off from the class they are
declared in, even if there are ordering rules for that class.

That is, even if `class {'update-repos': }` has ordering rules attached
to it, the actual code in the `update-repos` class could be executed
at any time without this anchor pattern being used.

This floating off is listed as a bug in Puppet, but the anchor pattern
is the current workaround. See:

http://projects.puppetlabs.com/issues/8040