Added module to install nginx.
This module installs nginx, manages its configuration file and starts
the daemon. This module is not meant to be extremely flexible. It does
not use templates for the configuration, but rather a simple
pre-written configuration file, loaded through a `file` directive.

Currently the only available configuration file is set up to host
static files and a gunicorn app.

This could potentially be expandable in the future to use other
directives, or to include multiple options for configuration files.